genai-function-calling/openai-agents/requirements.txt
openai-agents~=0.2.10
httpx~=0.28.1
mcp~=1.13.1

elastic-opentelemetry~=1.7.0
# Use openai-agents and MCP instrumentation from OpenInference
openinference-instrumentation-openai-agents~=1.2.0
# TODO: temporary until openinference-instrumentation-mcp 1.3.1 with ValidationError fix
openinference-instrumentation-mcp @ git+https://github.com/codefromthecrypt/openinference.git@mcp-validation-fix#subdirectory=python/instrumentation/openinference-instrumentation-mcp
